consistent naming verify_policy instead of verify-policy
